modules/szz_runner/requirements.txt
psycopg2-binary==2.9.9
PyYAML==6.0.2
pika==1.3.2
GitPython==3.1.43
PyDriller==1.15.5
dateparser==1.2.0
packaging==24.1
networkx==2.6.3
beautifulsoup4==4.10.0